+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `BM::Instrumentations::Process` collector to captures process' RSS memory and the number of open files
+
 ## [1.0.1] - 2021-05-29
 
 ### Added