+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changes
+
+- Update minimum puma version to `5.4.0`
+
 ## [1.0.1] - 2021-05-29
 
 ### Added