+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - `BM::Instrumentations::Process` collector to captures process' RSS memory and the number of open files
 
+### Changes
+
+- Update minimum puma version to `5.4.0`
+
 ## [1.0.1] - 2021-05-29
 
 ### Added